feat: 이미 참여한 챌린지일 경우 /my-poor-room으로 이동
